requirements.txt
Flask
Werkzeug
Pillow
pytesseract
python-dotenv
groq
gunicorn